Add files via upload
Completed on 10/09/2025, Wednesday. Uploaded on 11/09/2025 (I forgot to upload on the day… But I also got a version for today too). There were no classes today, but I had a two-hour workshop, so I wrote quite a lot of code and I’m almost finished.
Last Friday there was actually a class, but we also had to submit our algorithm on the same day and I was busy doing that. By the end of that day, I still didn’t manage to upload my code, and I hadn’t written much anyway, so I guess it wasn’t that important. On that Friday, since the Algorithm was due, I mainly focused on adding the functions into my code, testing for any logical errors, and then updating my flowchart. The functions I added are printing and showing images. That was basically all I did on Friday for the code.
On Wednesday, I thought that since my program had no layout at all, it would probably be difficult to read or understand, so I decided to finish the layout first before moving on to the final testing phase. That day, I started by fixing some bugs, not completely, but at least enough to make the program run. After that, I worked on the layout. It wasn’t really difficult, but it did take some time. Even after two hours, I still couldn’t finish it. The function of adding image had bugs too, which are waiting to be fixed.