Add files via upload
Completed on 12/09/2025, Friday. Uploaded on 16/09/2025, Tuesday.
 Why don’t the dates match again? I feel sorry, it wasn’t on purpose tho, I simply forgot (actually, on that Friday I originally planned to continue working on programming after school and upload then, but during the last class I thought, “it’s already Friday,” so in the end I decided not to go, and my code didn’t get uploaded).
The part I worked on the most that day was the images. Remember on Thursday I said I hadn’t finished the labels? What was missing were the labels for images, manage, and print. I finished the image and managed labels on Friday, and the print label was something I did today (Tuesday).
On Friday, I also added two buttons that let users choose images, which took a bit more time, but everything else was fine. I also fixed some bugs. At the end of the file, I listed a table of all the issues I could find that still need to be changed or fixed. These will be the final processes I need to clear once I finish the entire program.