Update EoY_v10.py
Completed and uploaded on Tuesday, 16/09/2025. Updated on 19/09/2025 for write my commit.
Today I finished my print function and applied it to the Print button and the Exit button, which triggered the final pop-up interface. I chose to use ttk.Treeview to present the results because it was the most convenient option. After that, I added a few small components that I hadn’t included before (such as a little monster displayed on the main interface), and in the process, I also fixed a few bugs. Judging from the results, I think I’ve completed everything in my code except for the parts that don’t run (e.g., delete card). I’m getting close to finishing, how nice.