Add files via upload
Completed on 18/09/2025, and uploaded the next day on 19/09/2025 because I was busy working on the code yesterday.
I fixed a large number of bugs, including those related to displayed_catalogue (I accidentally added a displayed_catalogue.clear() which caused a bug), error_popup (I had to use .lift() and .focus_force() on window["window2"] to make it appear in front of the main window), close_active_frame (frames were not being properly deleted), and many more.

I also made some improvements: the lines using image.open were too long, so I created a loop to simplify them. The update_catalogue function was added to my code. It does what its name suggests, updating and organizing the catalogue, and I added it throughout my program because it is very useful.

Since yesterday was the final code deadline, even though my program still has areas that could be improved and some remaining bugs, I could not continue working on it. Therefore, this is the last update of my code on GitHub, and this version is my final code.